requirements.txt
streamlit
plotly
pandas
requests
python-dotenv
openpyxl
